requirements.txt
# Core ML
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# Data processing
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2

# Experiment Tracking
mlflow==2.10.2

# DVC for data versioning
dvc==3.40.1

# REST API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# Monitoring
prometheus-client==0.20.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
